requirements.txt
blinker==1.6.2
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==2.0.0.post2
Flask==2.2.3
flask-mongoengine==1.0.0
Flask-PyMongo==2.3.0
Flask-WTF==1.1.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.3
mongoengine==0.27.0
numpy==1.25.0
pymongo==4.4.0
python-dotenv==1.0.0
scikit-learn==1.2.2
scipy==1.11.0
threadpoolctl==3.1.0
Werkzeug==2.3.6
WTForms==3.0.1